requirements.txt
aiohttp
psycopg2-binary